Changed to automatically return pounds. Currency now a settable value (in cookie being sent).

+++ scrape.js
@@ -1,7 +1,5 @@
-var XMLHttpRequest = require("xmlhttprequest").XMLHttpRequest;
 var cheerio = require('cheerio');
 var get = require('./config.json');
-var zlib = require('zlib');
 get.db.addTimestamp = true;
 var mysql = require('mysql');
 var fs = require('fs');
@@ -11,6 +9,7 @@ var timeStamp = 0;
 var thresh = 0;
 var absentConf = false;
 process.setMaxListeners(0);
+var callFinal = false;
 if (get.writeToDB == true) {
 	//	if (get.finalSearch.length != get.db.columns.length) {
 	//		console.log("search paramaters not equal to column parameters");
@@ -64,13 +63,11 @@ for (var i = diff.length - 1; i >= 0; i--) {
 	}
 }
 console.log("Time until repeat: " + timeStamp / 1000 / 60 + " minutes.");
-var connection = mysql.createConnection({
-	"host" : get.db.host,
-	"user" : get.db.user,
-	"password" : get.db.pass,
-	"database" : get.db.dbName
-
-});
+var connection = mysql.createConnection('mysql://'+get.db.user+':'+get.db.pass+'@'+get.db.host+'/'+get.db.dbName+'?reconnect=true');
+//	{"host" : get.db.host,
+//	"user" : get.db.user,
+//	"password" : get.db.pass,
+//	"database" : get.db.dbName}
 
 if (get.writeToDB == "true") {
 	//connection.query("TRUNCATE TABLE " + get.db.tName + " ;");
@@ -94,7 +91,13 @@ function timed() {
 	if (absentConf == true && get.recover == true) {
 		console.log("Didn't finish eating last time around!");
 		var encode = JSON.parse(fs.readFileSync("numTracker.json", "utf-8"));
-		writingToEndPoint(encode.data);
+		callFinal = !callFinal;
+		if(encode.data > 0){
+			writingToEndPoint(encode.data-1);
+		}
+		else{
+			writingToEndPoint(encode.data);
+		}
 	} else {
 		for (var j = 0; j < get.baseUrl.length; j++) {
 			huntLinks(j);
@@ -113,6 +116,7 @@ function huntLinks(j) {
 			carryOn(body, get.baseUrl[j], j);
 			console.log(get.baseUrl[j]);
 		} else {
+			console.log(response);
 			console.log("\033[31m A spider got lost! Error " + error + " with a code " + response.statusCode + '\033[0m');
 		}
 
@@ -173,11 +177,11 @@ function dataScrape(base, k, referenceArr) { {
 				//console.log(tempResults);
 				asyncLoop(tempResults.length - 1, function(loop) {
 					//				for (var l = 0; l < tempResults.length; l++) {
-					//var text = requester(tempResults[l]);
+					//var text =  requester(tempResults[l]);
 					//console.log(tempResults[loop.index]);
 					console.log(loop.index);
-					if(get.method == "asos"){
-						//tempResults[loop.index] = tempResults[loop.index]+"&parentID=-1&pge=0&pgeSize=9999&sort=-1";
+					if (get.method == "asos") {
+						tempResults[loop.index] = tempResults[loop.index]+"&parentID=-1&pge=0&pgeSize=9999&sort=-1";
 					}
 					request(tempResults[loop.index], function(error, response, body) {
 						if (!error && response.statusCode == 200) {
@@ -219,38 +223,31 @@ function dataScrape(base, k, referenceArr) { {
 							}
 							loop.next();
 							//console.log(resultsArr);
-							console.log('\033[35m' + Math.floor(100 * ((loop.index + 1) / tempResults.length)) + " % completed of pass " + (k + 1) + " of " + (get.searches.length - 1) + ". "+resultsArr.length+" results."+"\033[0m");
+							console.log('\033[35m' + Math.floor(100 * ((loop.index + 1) / tempResults.length)) + " % completed of pass " + (k + 1) + " of " + (get.searches.length - 1) + ". " + resultsArr.length + " results." + "\033[0m");
 
 						} else {
 							//console.log(tempResults);
-							console.log("\033[31m A spider got lost! Error " + error + " with a code " + response.statusCode + '\033[0m');
+							console.log("\033[31m A spider got lost! Error : " + error + '\033[0m');
 							loop.next();
 						}
 						if (loop.index == tempResults.length - 1) {
 							var pushk = k + 1;
-							dataScrape(base, pushk, resultsArr);
+							if(callFinal == true){
+								dataScrape(base, pushk, resultsArr);
+							}
+							callFinal = !callFinal;
 						}
 
 					})
 				}, function() {
 				});
 			} else if (k == get.searches.length - 1) {
-				//console.log(referenceArr);
 				console.log("finished");
-				//console.log(get.preserve);
-				// console.log(resultsArr);
-				//resultsArr = eliminateDuplicates(resultsArr);
-				//if (get.baseUrl.indexOf(base) > 0) {
-				//	var linkJson = JSON.parse(fs.readlinkSync('links.json'));
-				//			 fs.writeFileSync("links.json", '{"data":' + eliminateDuplicates(linkJson.data.concat(resultsArr)) + "}");
-				//} else {
-				//				console.log("writing this tree");
-				//					console.log(resultsArr);
 				fs.writeFileSync("links.json", '{"data":' + JSON.stringify(eliminateDuplicates(resultsArr)) + '}');
+				fs.writeFileSync("numTracker.json", '{"data":' + "0" + '}');
 				writingToEndPoint(0);
-
+				return;
 			}
-
 		} else if (get.method == "eat" && k == get.searches.length) {
 			if (get.writeToDB == "true") {
 				dropTable();
@@ -285,12 +282,20 @@ function writingToEndPoint(startFrom) {
 	console.log("file loaded");
 	//for (var l = 0; l < (links.length - startFrom); l++) {
 	var start = links.length - startFrom;
-	console.log(start);
+	console.log(startFrom);
+	console.log(links.length);
 	asyncLoop(start, function(loop) {
 		var l = loop.index;
-		console.log(links[l]);
-		console.log(l);
-		request(links[l], function(error, response, body) {
+		console.log(links[l+startFrom]);
+		console.log(l+startFrom);
+		var j = "";
+		if(get.method = "asos"){
+			var j = request.jar();
+			j.add(request.cookie("asos=topcatid=1000&customerguid=2a6aad1ef6d84192b3c634a4be54d846&PreferredSite=www.asos.com&currencyid=1"));
+		}
+		//request = request.defaults({jar:j});
+		request({url : links[l+startFrom],jar:j}, function(error, response, body) {
+			
 			if (!error && response.statusCode == 200) {
 				//console.log(body);
 				// console.log(body) // Print the web page.
@@ -299,90 +304,189 @@ function writingToEndPoint(startFrom) {
 				$ = cheerio.load(text);
 				var resultString = "";
 				var results = [];
-					
+
 				for (var m = 0; m < get.finalSearch.length; m++) {
 					var patternArr = parseElements(get.finalSearch[m]);
-					console.log(patternArr);
-											//price,imageUrl,title,category,description
+					//console.log(patternArr);
+					//price,imageUrl,title,category,description
 
 					for (var z = 0; z < patternArr.length; z++) {
-						if(get.method == "asos"){
-							if(get.finalSearch[m] == "#ctl00_ContentMainPage_ctlSeparateProduct_lblProductPrice"){
-								var temp=$(patternArr[z].replace(/>/g, " "));
+						if (get.method == "asos") {
+							if (get.finalSearch[m] == "#ctl00_ContentMainPage_divmor>ul>li>a"){
+								var temp = $(patternArr[z].replace(/>/g, " "));
+								if(temp.length>0){
+								//console.log(temp[0].children[0].data);
+								results.push(temp[0].children[0].data);
+								}
+								else{
+								results.push("null");
+								}
+							}
+							if (get.finalSearch[m] == "#ctl00_ContentMainPage_ctlSeparateProduct_lblProductPrice") {
+								var temp = $(patternArr[z].replace(/>/g, " "));
 								//console.log(temp[0].children[0].data.substring(6,temp[0].children[0].data.length));
-								results.push(temp[0].children[0].data.substring(6,temp[0].children[0].data.length));
-							} 
-							if(get.finalSearch[m] == "#ctl00_ContentMainPage_imgMainImage"){
-								var temp=$(patternArr[z].replace(/>/g, " "));
+								if(temp!=null){
+									if(temp.length == 0){
+										results.push("null");
+									}
+									else{
+										//console.log(temp[0].children[0]);
+										results.push(temp[0].children[0].data.substring(6, temp[0].children[0].data.length));
+									}
+								}
+								else{
+								results.push("null");
+							}
+							}
+							if (get.finalSearch[m] == "#ctl00_ContentMainPage_imgMainImage") {
+								var temp = $(patternArr[z].replace(/>/g, " "));
 								//console.log(temp[0].attribs.src);
-								results.push(temp[0].attribs.src);
-							}
-							if(get.finalSearch[m] == "#ctl00_ContentMainPage_ctlSeparateProduct_lblProductTitle"){
-								
-								var temp=$(patternArr[z].replace(/>/g, " "));
-								//console.log(temp[0].children);
-								var ans = "";
-								for(var i in temp[0].children){
-									if(temp[0].children[i].type == "text"){
-										if(ans.length > 0){ans+=","}
-										ans += temp[0].children[i].data;
+								if(temp!=null){
+									if(temp.length == 0){
+										results.push("null");
+									}
+									else{
+										results.push(temp[0].attribs.src);
 									}
 								}
-								results.push(ans);
+								else{
+								results.push("null");
 							}
-							if(get.finalSearch[m] == ".single-entry"){
-								var temp2=$(patternArr[z].replace(/>/g, " "))
-								var string = "";
-								for(var i in temp2[0].children){
-									if(temp2[0].children[i].name=='a'){
-										string = temp2[0].children[i].attribs.href;
-										break;		
-									}
-								}
-								//console.log(string.split("/")[2]);
-								results.push(string.split("/")[2]);
-								//console.log(")__________________________________()")
+							}
+							if (get.finalSearch[m] == "#ctl00_ContentMainPage_ctlSeparateProduct_lblProductTitle") {
 
-								var temp=$(patternArr[z].replace(/>/g, " ")).html();
-								//console.log(temp);
-								results.push(temp);
+								var temp = $(patternArr[z].replace(/>/g, " "));
+								//console.log(temp[0].children);
+								if(temp!=null){
+									if(temp.length == 0){
+										results.push("null");
+									}
+									else{
+									var ans = "";
+									for (var i in temp[0].children) {
+										if (temp[0].children[i].type == "text") {
+											if (ans.length > 0) {
+												ans += ","
+											}
+											ans += temp[0].children[i].data;
+										}
+									}
+									results.push(ans);
+								}
+								}
+								else{
+								results.push("null");
+							}
+							}
+							if (get.finalSearch[m] == "#ctl00_ContentMainPage_divmor>ul>li"){
+								var temp2 = $(patternArr[z].replace(/>/g, " "))
+								if (temp2 != null) {
+									if (temp2.length == 0) {
+										results.push("null");
+									} else {
+										logger = [];
+										for(var i = 0 ; i < temp2.length; i++){
+											logger.push(temp2[i].children[4].children[0].data);
+//											if(string.length > 0){string+="/";}
+	//										console.log(temp2[i].children[4].children[0].data);											
+		//									string+=temp2[i].children[4].children[0].data;
+										}
+										//console.log(logger);
+										if(logger.length == 0){
+											results.push("null");
+										}
+										else{
+											logger = eliminateDuplicates(logger);
+											//console.log(logger);
+											var string = "";
+											for(var i in logger){
+												if(string.length>0){
+													string += "/";
+												}
+												string+= logger[i];
+											}
+											//console.log(string);
+											results.push(string);
+										}
+									}
+								} else {
+									results.push("null");
+								}
+	
+							}
+							if (get.finalSearch[m] == ".single-entry") {
+							
+								var temp = $(patternArr[z].replace(/>/g, " ")).html();
+									if(temp!=null){
+										if(temp.length == 0){
+											results.push("null");
+										}
+										else{
+										//console.log(temp);
+										results.push(temp);
+									}
+								}
+								else{
+								results.push("null");
+							}
 							}
 						}
-						if(get.method == "generic"){
+						if (get.method == "generic") {
 							var temp = $(patternArr[z].replace(/>/g, " "));
 							//console.log($._root.children);
-							for (var y = 0; y < temp.length; y++) {
-								results.push(temp[y]);
+							if(temp!=null){	
+								for (var y = 0; y < temp.length; y++) {
+									results.push(temp[y]);
+								}
+							}
+							else{
+								results.push("null");
 							}
 						}
 					}
+				}
 					//console.log(results.length);
+					if(get.method == "asos"){
+						if(results.length!=6){
+							console.log(results);
+							console.log("Scraper didn't retrieve results. X_X.");
+							process.exit(1);
+						}	
+					}
+
 					//results = eliminateDuplicates(results);
 					//console.log(results[0].children);
 					for (var n = 0; n < results.length; n++) {
 						//console.log(results[n]);
 						//console.log(results[n].children[0].data);
-						
-						if(get.method == "generic"){
+						if(get.method == "asos"){
+							if(resultString.length > 0){
+								resultString+=",";
+							}
+							resultString+=connection.escape(results[n]);
+							if(n == results.length-1){
+								resultString+=","+connection.escape(links[l+startFrom]);
+							}
+						}
+						if (get.method == "generic") {
 							if ("children" in results[n] && results[n].children.length > 0) {
 								if ("data" in results[n].children[0]) {
-								
-										if (resultString.length > 0 && resultString[resultString.length - 1] != ",") {
-											resultString += ",";
-										}
-										resultString += results[n].children[0].data;
-										console.log(results[n].children[0].data);
-									
+
+									if (resultString.length > 0 && resultString[resultString.length - 1] != ",") {
+										resultString += ",";
+									}
+									resultString += results[n].children[0].data;
+									console.log(results[n].children[0].data);
+
 								}
 							}
 						}
 					}
-				}
-				console.log(results);
-				console.log(results.length);
-//				console.log(resultString);
+				
+				//console.log("\033[1;35m" +results+"\033[0m");
+				//console.log(results.length);
+								//console.log(resultString);
 				if (get.writeToDB == true) {
-					console.log("happening");
 					if (resultString != "") {
 						//	writeToDb(connection.escape(resultString) + "," + connection.escape(links[l + startFrom]));
 
@@ -395,14 +499,15 @@ function writingToEndPoint(startFrom) {
 							}
 							string += get.db.columns[i];
 						}
-						//console.log(string);
-						resultString = connection.escape(resultString)+","+connection.escape(links[loop.index]);
+						console.log(string);
+						if(get.method=="generic"){
+							resultString = connection.escape(resultString) + "," + connection.escape(links[loop.index]);
+						}
 						if (get.db.addTimestamp == true) {
 							var sql = "INSERT INTO " + get.db.tName + " (" + string + ",timestamp) VALUES (" + resultString + "," + new Date().getTime() + ")";
 						} else {
-							var sql = "INSERT INTO " + get.db.tName + " (" + string + ") VALUES (" + resultString +")";
+							var sql = "INSERT INTO " + get.db.tName + " (" + string + ") VALUES (" + resultString + ")";
 						}
-
 						sql += " ON DUPLICATE KEY UPDATE ";
 						var dataSplit = resultString.replace(/" "/g, "").replace(/','/g, "'|'");
 						for (var z = 0; z < string.split(",").length; z++) {
@@ -415,30 +520,43 @@ function writingToEndPoint(startFrom) {
 						if (get.db.addTimestamp == true) {
 							sql += ", timestamp = " + new Date().getTime();
 						}
-						console.log("\033[31m" + sql + "\033[0m" + "\n");
-
+						//console.log(resultString);
+						//console.log("\033[31m" + sql + "\033[0m" + "\n");
 						connection.query(sql, function(err, results) {
-							if (err != undefined) {
-								console.log(err);
-								return;
-							}
 							console.log(results);
+							if (err != undefined) {
+								console.log("\033[1;35m" +err+"\033[0m");
+								//return;
+							}
+							//console.log(results);
+							fs.writeFileSync("numTracker.json", '{"data":' + (l + startFrom) + '}');
+							console.log("Wrapping in silk at \033[1;35m" + Math.floor(100 * ((l + startFrom) / links.length)) + "%\033[0m completion. \033[1;35m" + (links.length - (l + startFrom)) + "\033[0m links remaining.");
+							loop.next();
 						});
 						//loop.next();
 
 					}
 					//writeToDB();
 				}
+				else{
 				fs.writeFileSync("numTracker.json", '{"data":' + (l + startFrom) + '}');
-				console.log("Wrapping in silk at " + Math.floor(100 * ((l + startFrom) / links.length)) + "% completion. " + (links.length - (l + startFrom)) + " links remaining.");
+				console.log("Wrapping in silk at \033[1;35m" + Math.floor(100 * ((l + startFrom) / links.length)) + "%\033[0m completion. \033[1;35m" + (links.length - (l + startFrom)) + "\033[0m links remaining.");
+				//console.log("\033[1;35m" +resultString+"\033[0m");
+				loop.next();
+				}
 			} else {
-				console.log("A spider got lost! Error " + error + " with a code " + response.statusCode);
+				console.log("\033[31mA spider got lost! Error " + error+"\033[0m");
+				loop.next();
 			}
-			loop.next();
+			//loop.next();
 		})
 	}, function() {
-	fs.unlinkSync('numTracker.json');
-	fs.unlinkSync('links.json');
+		callFinal = !callFinal;
+		console.log(fs.existsSync('numTracker.json')+","+fs.existsSync('links.json'));
+		if(fs.existsSync('numTracker.json')){
+			fs.unlinkSync('numTracker.json');
+		}
+		
 	});
 }
